+++ src/shared/inputs/SingleInput.h
@@ -92,12 +92,13 @@ public:
         /* Mouse */
         IT_POINTER_X = 1,
         IT_POINTER_Y = 2,
-        IT_POINTER_MODE = 3,
-        IT_POINTER_B1 = 4 + POINTER_B1,
-        IT_POINTER_B2 = 4 + POINTER_B2,
-        IT_POINTER_B3 = 4 + POINTER_B3,
-        IT_POINTER_B4 = 4 + POINTER_B4,
-        IT_POINTER_B5 = 4 + POINTER_B5,
+        IT_POINTER_WHEEL = 3,
+        IT_POINTER_MODE = 4,
+        IT_POINTER_B1 = 5 + POINTER_B1,
+        IT_POINTER_B2 = 5 + POINTER_B2,
+        IT_POINTER_B3 = 5 + POINTER_B3,
+        IT_POINTER_B4 = 5 + POINTER_B4,
+        IT_POINTER_B5 = 5 + POINTER_B5,
 
         /* Single flag */
         IT_FLAG = 10,